+++ src/lib/lisp/repl.c
@@ -29,10 +29,61 @@ lispptr current_toplevel;
 char    num_repls;          // REPL count.
 #ifndef NO_DEBUGGER
 char    num_debugger_repls; // Debugger REPL count.
+lispptr debugger_return_value_sym; // Old (erroraneous) 'value'.
 #endif
 bool    do_break_repl;      // Tells current REPL to return.
 bool    do_continue_repl;   // If do_break_repl, tell REPL to continue.
 bool    do_exit_program;    // Return to top-level REPL.
+
+void
+out_colon (void)
+{
+    outs (": ");
+}
+
+void
+print_debugger_info ()
+{
+    if (error_code) {
+        outs ("Error #");
+        outn (error_code);
+        out_colon ();
+        if (last_errstr)
+            outs (last_errstr);
+        terpri ();
+    }
+    fresh_line ();
+
+    do_highlight = true;
+#ifdef DEBUG_INTERNAL
+    outs ("Eval'd: ");
+    print (current_expr);
+    terpri ();
+#endif
+
+    // Print last result unless it's the current expression.
+    if (value != current_function && value != current_toplevel) {
+        outs ("Rvalue: ");
+        print (value);
+        terpri ();
+    }
+
+    outs ("In");
+    if (current_function) {
+        tmp2 = SYMBOL_VALUE(current_function);
+        print (current_function);
+        out (' ');
+        print (FUNARGS(tmp2));
+        out_colon ();
+        terpri ();
+        print (FUNBODY(tmp2));
+    } else {
+        out_colon ();
+        print (current_toplevel);
+    }
+    do_highlight = false;
+    terpri ();
+}
 
 lispptr FASTCALL
 lisp_repl (char mode)
@@ -56,6 +107,7 @@ lisp_repl (char mode)
     num_repls++;
 #ifndef NO_DEBUGGER
     if (mode == REPL_DEBUGGER) {
+        SET_SYMBOL_VALUE(debugger_return_value_sym, value);
         num_debugger_repls++;
         outs ("In debugger #");
         outn (num_debugger_repls);
@@ -89,7 +141,7 @@ lisp_repl (char mode)
 
 #ifdef NO_DEBUGGER
         // Error not handled.  Exit program.
-        print_error_info ();
+        print_debugger_info ();
         do_break_repl   = true;
         do_exit_program = true;
         error_code      = 0;
@@ -104,7 +156,7 @@ lisp_repl (char mode)
     while (!eof ()) {
 #ifndef NO_DEBUGGER
         if (mode == REPL_DEBUGGER) {
-            print_error_info ();
+            print_debugger_info ();
             error_code = 0;
         }
 #endif
@@ -151,6 +203,8 @@ lisp_repl (char mode)
 
                 // Print expression.
                 case 'p':
+                    PUSH(SYMBOL_VALUE(debugger_return_value_sym));
+                    PUSH(value);
                     PUSH(x);
                     x = read ();
                     terpri ();
@@ -161,6 +215,9 @@ lisp_repl (char mode)
                     print (tmp);
                     tmp = nil;
                     terpri ();
+                    POP(value);
+                    POP(tmp);
+                    SET_SYMBOL_VALUE(debugger_return_value_sym, tmp);
                     goto next;
 
                 default:
@@ -185,12 +242,11 @@ lisp_repl (char mode)
 
         // Macro expansion if MACROEXPAND is a user function.
         if (CONSP(SYMBOL_VALUE(macroexpand_sym))) {
-#ifdef VERBOSE_MACROEXPAND
-            fresh_line (); outs ("X ");
-#endif
+#ifndef NO_DEBUGGER
             // Avoid debug step into MACROEXPAND.
             PUSH(debug_step);
             debug_step = nil;
+#endif
 
             // Call MACROEXPAND.
             x = make_cons (x, nil);
@@ -199,7 +255,9 @@ lisp_repl (char mode)
             PUSH_TAG(TAG_DONE);
             x = eval0 ();
 
+#ifndef NO_DEBUGGER
             POP(debug_step);
+#endif
         }
 #endif // #ifndef NO_MACROEXPAND
 
@@ -336,6 +394,8 @@ init_repl ()
     num_repls          = 0;
 #ifndef NO_DEBUGGER
     num_debugger_repls = 0;
+    debugger_return_value_sym = make_symbol ("*r*", 3);
+    expand_universe (debugger_return_value_sym);
 #endif
 #ifndef NO_MACROEXPAND
     unexpanded_toplevel = nil;