+++ Changelog.md
@@ -27,10 +27,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add PARANOID relocation table overflow check to GC sweep phase.
 - Option VERBOSE\_COMPRESSED\_CONS for diagnostic printing a 'C' for
   each compressed cons.
+- Breakpoints on user-defined procedures. (Symbol list "\*b\*".)
+- Debugger keeps its current return value in symbol "\*r\*".
 
 #### Changed
 
 - Interpreter exits if out of heap instead of going vodka.  This is temporary.
+- Debugger command 'p': Does not modify the return value.
 
 ### libsimpleio